apps/api/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
drf-spectacular>=0.28,<0.29
django-cors-headers>=4.7,<4.8
psycopg[binary]>=3.2,<3.3
celery>=5.5,<5.6
redis>=5.2,<6.0

gunicorn>=23.0,<24.0
whitenoise>=6.8,<7.0
